+++ src/unistr.cpp
@@ -101,7 +101,9 @@ void init_unistr(py::module &m, py::class_<Replaceable, UObject> &rep, py::class
 
   us.def(
       "__ge__",
-      [](const UnicodeString &self, const _UnicodeStringVariant &other) { return self >= VARIANT_TO_UNISTR(other); },
+      [](const UnicodeString &self, const _UnicodeStringVariant &other) -> bool {
+        return self >= VARIANT_TO_UNISTR(other);
+      },
       py::is_operator(), py::arg("other"));
 
   us.def(
@@ -135,7 +137,9 @@ void init_unistr(py::module &m, py::class_<Replaceable, UObject> &rep, py::class
 
   us.def(
       "__gt__",
-      [](const UnicodeString &self, const _UnicodeStringVariant &other) { return self > VARIANT_TO_UNISTR(other); },
+      [](const UnicodeString &self, const _UnicodeStringVariant &other) -> bool {
+        return self > VARIANT_TO_UNISTR(other);
+      },
       py::is_operator(), py::arg("other"));
 
   us.def("__hash__", &UnicodeString::hashCode);
@@ -151,14 +155,18 @@ void init_unistr(py::module &m, py::class_<Replaceable, UObject> &rep, py::class
 
   us.def(
       "__le__",
-      [](const UnicodeString &self, const _UnicodeStringVariant &other) { return self <= VARIANT_TO_UNISTR(other); },
+      [](const UnicodeString &self, const _UnicodeStringVariant &other) -> bool {
+        return self <= VARIANT_TO_UNISTR(other);
+      },
       py::is_operator(), py::arg("other"));
 
   us.def("__len__", &UnicodeString::length);
 
   us.def(
       "__lt__",
-      [](const UnicodeString &self, const _UnicodeStringVariant &other) { return self < VARIANT_TO_UNISTR(other); },
+      [](const UnicodeString &self, const _UnicodeStringVariant &other) -> bool {
+        return self < VARIANT_TO_UNISTR(other);
+      },
       py::is_operator(), py::arg("other"));
 
   us.def(