+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - [FIXED] Raise IndexError when \_ConstChar16Ptr.\_\_getitem__() accessing beyond capacity
 - [FIXED] Change return type of UnicodeSet.\_\_getitem__(index: slice) -> list[int] → UnicodeSet
 - [FIXED] Change return type of UnicodeString.\_\_getitem__(index: slice) -> str → UnicodeString
+- [FIXED] Change return type of UnicodeString comparison operators from int to bool
 - [FIXED] Change UTextVector to keep references
 - [REMOVED] Remove UTextVector(n: int)